### 2018-07-24 #835 @whedon
[PRE REVIEW]: UMAP: Uniform Manifold Approximation and Projection
**Submitting author:** @lmcinnes (<a href="http://orcid.org/0000-0003-2143-6834">Leland McInnes</a>)
**Repository:** <a href="https://github.com/lmcinnes/umap" target ="_blank">https://github.com/lmcinnes/umap</a>
**Version:** v0.3.0
