### 2018-07-25 9f9870d
Fix fit(), then transform() bug referenced in #89
