### 2018-07-26 #835 @arokem
Happy to do it! I saw the scipy talk on youtube and enjoyed that. Happy to see this paper submitted for review in JOSS.

@lmcinnes: do you have suggestions for reviewers? You can see a list of potential reviewers [here](https://bit.ly/joss-reviewers).
