### 2018-07-26 #835 @lmcinnes
Glad you enjoyed the talk! That is an impressively long list of reviewers. Some potential suggestions based on that list are: glemaitre; stsievert; jsgalan; rabst; terrytangyuan, although there are a number of other clearly qualified individuals on that list, so I would also be happy with many others. I would also note that corybrunson has ideal expertise (topological data analysis in particular), but only lists R. If they were willing they would also make a great reviewer.
