### 2018-07-27 #835 @arokem
👋 @terrytangyuan : would you be willing to review this paper for JOSS?
