### 2018-07-29 cbf1709
add the ability to specify a non-euclidean metric for the embedding space
