+++ umap/umap_.py
@@ -18,7 +18,7 @@ import scipy.sparse
 import scipy.sparse.csgraph
 import numba
 
-import umap.distances as dist
+import umap.distances
 
 import umap.sparse as sparse
 
@@ -195,8 +195,8 @@ def nearest_neighbors(
     else:
         if callable(metric):
             distance_func = metric
-        elif metric in dist.named_distances:
-            distance_func = dist.named_distances[metric]
+        elif metric in umap.distances.named_distances:
+            distance_func = umap.distances.named_distances[metric]
         else:
             raise ValueError("Metric is neither callable, " + "nor a recognised string")
 
@@ -677,147 +677,152 @@ def rdist(x, y):
     return result
 
 
-@numba.njit(fastmath=True)
-def optimize_layout(
-    head_embedding,
-    tail_embedding,
-    head,
-    tail,
-    n_epochs,
-    n_vertices,
-    epochs_per_sample,
-    a,
-    b,
-    rng_state,
-    gamma=1.0,
-    initial_alpha=1.0,
-    negative_sample_rate=5.0,
-    verbose=False,
+def make_optimize_layout(
+    output_metric,
+    output_metric_kwds,
 ):
-    """Improve an embedding using stochastic gradient descent to minimize the
-    fuzzy set cross entropy between the 1-skeletons of the high dimensional
-    and low dimensional fuzzy simplicial sets. In practice this is done by
-    sampling edges based on their membership strength (with the (1-p) terms
-    coming from negative sampling similar to word2vec).
+    @numba.njit(fastmath=True)
+    def optimize_layout(
+        head_embedding,
+        tail_embedding,
+        head,
+        tail,
+        n_epochs,
+        n_vertices,
+        epochs_per_sample,
+        a,
+        b,
+        rng_state,
+        gamma=1.0,
+        initial_alpha=1.0,
+        negative_sample_rate=5.0,
+        verbose=False,
+    ):
+        """Improve an embedding using stochastic gradient descent to minimize the
+        fuzzy set cross entropy between the 1-skeletons of the high dimensional
+        and low dimensional fuzzy simplicial sets. In practice this is done by
+        sampling edges based on their membership strength (with the (1-p) terms
+        coming from negative sampling similar to word2vec).
 
-    Parameters
-    ----------
-    head_embedding: array of shape (n_samples, n_components)
-        The initial embedding to be improved by SGD.
+        Parameters
+        ----------
+        head_embedding: array of shape (n_samples, n_components)
+            The initial embedding to be improved by SGD.
 
-    tail_embedding: array of shape (source_samples, n_components)
-        The reference embedding of embedded points. If not embedding new
-        previously unseen points with respect to an existing embedding this
-        is simply the head_embedding (again); otherwise it provides the
-        existing embedding to embed with respect to.
+        tail_embedding: array of shape (source_samples, n_components)
+            The reference embedding of embedded points. If not embedding new
+            previously unseen points with respect to an existing embedding this
+            is simply the head_embedding (again); otherwise it provides the
+            existing embedding to embed with respect to.
 
-    head: array of shape (n_1_simplices)
-        The indices of the heads of 1-simplices with non-zero membership.
+        head: array of shape (n_1_simplices)
+            The indices of the heads of 1-simplices with non-zero membership.
 
-    tail: array of shape (n_1_simplices)
-        The indices of the tails of 1-simplices with non-zero membership.
+        tail: array of shape (n_1_simplices)
+            The indices of the tails of 1-simplices with non-zero membership.
 
-    n_epochs: int
-        The number of training epochs to use in optimization.
+        n_epochs: int
+            The number of training epochs to use in optimization.
 
-    n_vertices: int
[…]
