### 2018-07-31 781ff33
Fix optimize_layout so it actually uses \frac{dD(x,y)}{dx} in its calculation. Added functions for euclidean and haversine metrics that return the distance and the gradient wrt the first point. Added lookup table for gradient functions to distances.py and logic to UMAP class to pass the gradient function to make_optimize_layout.
