+++ umap/umap_.py
@@ -678,9 +678,30 @@ def rdist(x, y):
 
 
 def make_optimize_layout(
-    output_metric,
+    output_metric_grad,
     output_metric_kwds,
 ):
+    """Create a numba accelerated function to improve an embedding using
+    stochastic gradient descent to minimize the fuzzy set cross entropy
+    between the 1-skeletons of the high dimensional and low dimensional
+    fuzzy simplicial sets. In practice this is done by sampling edges
+    based on their membership strength (with the (1-p) terms coming from
+    negative sampling similar to word2vec).
+
+    Parameters
+    ----------
+    output_metric_grad: function
+        Function returning the distance between two points in embedding
+        space and the gradient of the distance wrt the first argument.
+
+    output_metric_kwds: dict
+        Key word arguments to be passed to the output_metric_grad function
+
+    Returns
+    -------
+    A numba JITd function for optimizing an embeddinng that is specialised
+    to the given metric in the embedding space.
+    """
     @numba.njit(fastmath=True)
     def optimize_layout(
         head_embedding,
@@ -775,13 +796,17 @@ def make_optimize_layout(
                     current = head_embedding[j]
                     other = tail_embedding[k]
 
-                    dist_squared = output_metric(current, other, *output_metric_kwds)**2
+                    dist_output, grad_dist_output = output_metric_grad(
+                        current, other,*output_metric_kwds
+                    )
 
-                    grad_coeff = -2.0 * a * b * pow(dist_squared, b - 1.0)
-                    grad_coeff /= a * pow(dist_squared, b) + 1.0
+                    grad_coeff = -2.0 * a * b * pow(dist_output, 2*b - 1)
+                    grad_coeff /= (a * pow(dist_output, 2*b) + 1.0)
 
                     for d in range(dim):
-                        grad_d = clip(grad_coeff * (current[d] - other[d]))
+                        grad_d = clip(grad_coeff * grad_dist_output[d])
+                        # grad_d = clip(grad_coeff * (current[d] - other[d]) / (0.001 + dist_output))
+
                         current[d] += grad_d * alpha
                         if move_other:
                             other[d] += -grad_d * alpha
@@ -798,18 +823,21 @@ def make_optimize_layout(
 
                         other = tail_embedding[k]
 
-                        dist_squared = output_metric(current, other, *output_metric_kwds)**2
+                        dist_output, grad_dist_output = output_metric_grad(
+                            current, other, *output_metric_kwds
+                        )
 
                         grad_coeff = 2.0 * gamma * b
-                        grad_coeff /= (0.001 + dist_squared) * (
-                            a * pow(dist_squared, b) + 1
+                        grad_coeff /= (0.001 + dist_output) * (
+                            a * pow(dist_output, 2*b) + 1
                         )
 
                         if not np.isfinite(grad_coeff):
                             grad_coeff = 4.0
 
                         for d in range(dim):
-                            grad_d = clip(grad_coeff * (current[d] - other[d]))
+                            grad_d = clip(grad_coeff * grad_dist_output[d])
+                            # grad_d = clip(grad_coeff * (current[d] - other[d]) / (0.001 + dist_output))
                             current[d] += grad_d * alpha
 
                     epoch_of_next_negative_sample[i] += (
@@ -839,7 +867,7 @@ def simplicial_set_embedding(
     random_state,
     input_metric,
     input_metric_kwds,
-    output_metric,
+    output_metric_grad,
     output_metric_kwds,
     verbose,
 ):
@@ -894,13 +922,20 @@ def simplicial_set_embedding(
     random_state: numpy RandomState or equivalent
         A state capable being used as a numpy random state.
 
-    input_metric: string
[…]
