### 2018-07-31 62b8bc9
Add example for embedding onto a torus and a sphere. Fix argument checking for output_metric_grad
