### 2018-08-01 8afe0f9
Fix issue #98. Bandwidth got deprecated.
