### 2018-08-01 896a17f
Added inverse option to optimize_layout to allow the use of the data space membership function rather than the embedding space membership function. The gradient of the cross entropy stays the same because the positions of w_h and w_l swap since we are now taking the embedding space to be ground truth and the expectation should be taken with respect to the embedding edge probabilities.
