### 2018-08-02 c46adeb
Fix typo in JOSS paper
