### 2018-08-02 #861 @lmcinnes
Fixed a typo in paper.md. Otherwise all looks good.
