### 2018-08-02 bf77f91
Typo in correction to local_connectivity for transform
