### 2018-08-04 2870a4e
Implement inverse_transform using a Delaunay commplex rather than the Vietoris-Rips complex used in the forward transform. This gives intuitive, smooth behavior in gaps between clusters but means that transform(inverse_transform(pts)) != pts in general. Also added example script in exaples directory.
