+++ umap/umap_.py
@@ -43,6 +43,37 @@ SMOOTH_K_TOLERANCE = 1e-5
 MIN_K_DIST_SCALE = 1e-3
 NPY_INFINITY = np.inf
 
+def bfs(adjmat, start, min_vertices):
+    # keep track of all visited nodes
+    explored = []
+    # keep track of nodes to be checked
+    queue = [start]
+ 
+    levels = {}         # this dict keeps track of levels
+    levels[start] = 0    # depth of start node is 0
+    max_level = np.inf
+ 
+    visited = [start]     # to avoid inserting the same node twice into the queue
+ 
+    # keep looping until there are nodes still to be checked
+    while queue:
+        # pop shallowest node (first node) from queue
+        node = queue.pop(0)
+        explored.append(node)
+        if max_level == np.inf and len(explored) > min_vertices:
+            max_level = max(levels.values())
+ 
+        if levels[node] + 1 < max_level:
+            neighbors = adjmat[node].indices
+            # add neighbors of node to queue
+            for neighbour in neighbors:
+                if neighbour not in visited:
+                    queue.append(neighbour)
+                    visited.append(neighbour)
+ 
+                    levels[neighbour] = levels[node] + 1
+ 
+    return np.array(explored)
 
 @numba.njit(parallel=True, fastmath=True)
 def smooth_knn_dist(distances, k, n_iter=64, local_connectivity=1.0, bandwidth=1.0):
@@ -678,7 +709,7 @@ def rdist(x, y):
 
 
 def make_optimize_layout(
-    output_metric_grad,
+    output_metric,
     output_metric_kwds,
 ):
     """Create a numba accelerated function to improve an embedding using
@@ -690,12 +721,12 @@ def make_optimize_layout(
 
     Parameters
     ----------
-    output_metric_grad: function
+    output_metric: function
         Function returning the distance between two points in embedding
         space and the gradient of the distance wrt the first argument.
 
     output_metric_kwds: dict
-        Key word arguments to be passed to the output_metric_grad function
+        Key word arguments to be passed to the output_metric function
 
     Returns
     -------
@@ -720,7 +751,7 @@ def make_optimize_layout(
         initial_alpha=1.0,
         negative_sample_rate=5.0,
         verbose=False,
-        inverse=False
+        inverse=False,
     ):
         """Improve an embedding using stochastic gradient descent to minimize the
         fuzzy set cross entropy between the 1-skeletons of the high dimensional
@@ -803,15 +834,18 @@ def make_optimize_layout(
                     current = head_embedding[j]
                     other = tail_embedding[k]
 
-                    dist_output, grad_dist_output = output_metric_grad(
+                    dist_output, grad_dist_output = output_metric(
                         current, other, *output_metric_kwds
                     )
 
                     w_l = pow((1 + a * pow(dist_output, 2 * b)), -1)
                     if inverse:
-                        grad_coeff = (w_l - 1) / ((1 - w_h) * sigmas[i])
+                        grad_coeff = (
+                            (w_l - 1) /
+                            ((1 - w_h) * sigmas[j] + 1e-6)
+                        )
                     else:
-                        grad_coeff = 2 * b * (w_l - 1)/(dist_output + 0.001)
+                        grad_coeff = 2 * b * (w_l - 1)/(dist_output + 1e-6)
 
                     for d in range(dim):
                         grad_d = clip(grad_coeff * grad_dist_output[d])
@@ -829,18 +863,27 @@ def make_optimize_layout(
 
                     for p in range(n_neg_samples):
                         k = tau_rand_int(rng_state) % n_vertices
+                        # for negative samples, the edge does not exist so
+                        # w_h == 0.0
+                        w_h = 0.0
 
                         other = tail_embedding[k]
 
[…]
