+++ examples/inverse_transform_example.py
@@ -14,7 +14,7 @@ X_train, X_test, y_train, y_test = train_test_split(
     random_state=42
 )
 
-trans = umap.UMAP(s
+trans = umap.UMAP(
     n_neighbors=10,
     random_state=42,
     input_metric='euclidean',
@@ -37,16 +37,16 @@ test_pts = np.array([
     for x in np.linspace(0, 1, 10)
 ])
 
-inv_trasnformed_points = trans.inverse_transform(test_pts)
+inv_transformed_points = trans.inverse_transform(test_pts)
 
 plt.scatter(trans.embedding_[:, 0], trans.embedding_[:, 1], c=y_train, cmap='tab10')
 plt.scatter(test_pts[:, 0], test_pts[:, 1], marker='x', c='k')
 
-plt.figure()
-fig, ax = plt.subplots(10, 10, squeeze=True)
-for i in range(100):
-    ax[i].imshow(inv_trasnformed_points[i].reshape(28, 28), origin='upper')
-    ax[i].get_xaxis().set_visible(False)
-    ax[i].get_yaxis().set_visible(False)
+fig, ax = plt.subplots(10, 10)
+for i in range(10):
+    for j in range(10):
+        ax[i,j].imshow(inv_transformed_points[i*10+j].reshape(28, 28), origin='upper')
+        ax[i,j].get_xaxis().set_visible(False)
+        ax[i,j].get_yaxis().set_visible(False)
 
 plt.show()
