### 2018-08-06 d286224
Clarify warning as per #108
