### 2018-08-07 c7aa43d
Fix issue #107
