### 2018-08-08 3dcf706
Fix for issue #102
