### 2018-08-08 aa59fee
Fix for issue #112
