### 2018-08-09 7a0c49d
Attempt to fix issue #110 at the source.
