### 2018-08-09 02afd05
Add comment clarifying an issue raised in PR #103
