### 2018-08-10 8e81bfb
Update umap_.py
