### 2018-08-13 866c079
Fix issue #85; easy to fix, hard to find.
