### 2018-08-17 7e490b4
Better error message per issue #124
