### 2018-08-21 #861 @terrytangyuan
Hi @lmcinnes, thanks for your ping! It is a well-written package and has already earned 1.5k stars.

Two things that are missing:

* References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)? You need at list one reference with DOI.
* Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Besides, a couple suggestions:

* It might be worthwhile showing examples to use UMAP together with sklearn's pipelines, e.g. after UNMAP transformation, apply an sklearn estimator to show the differences in performance.
* Considering providing utility functions to plot the result of UNMAP, e.g. example [here](https://github.com/lmcinnes/umap/blob/97d33f57459de796774ab2d7fcf73c639835676d/examples/plot_algorithm_comparison.py) might be too verbose and hard-to-understand for beginners.
