### 2018-08-22 #861 @lmcinnes
Hi @terrytangyuan ! Thanks for the feedback. I have added a CONTRIBUTING.md to provide some guidelines of contributions which goes into more detail than the small "Contributing" section in the README.rst. The project already has a CODE_OF_CONDUCT.md that should cover community behaviour when contributing.

The references issue is more difficult. Currently UMAP is still a pre-print, and arXiv does not provide DOIs. As that pre-print is currently the only reference this makes things difficult. Is there any wiggle room on the requirement for a DOI for a reference? I can certainly ensure the code, when it gets archived for inclusion, gets a DOI, but that isn't a reference. I would welcome any thoughts on the right approach to this.

Finally, thanks for the suggestions! Both are excellent, and I'll try to get something sorted out soon -- they will take a little more work.
