### 2018-08-22 lmcinnes/umap#127 @lmcinnes
Add plotting functions
As per @terrytangyuan in openjournals/joss-reviews#861 , it would be helpful to have an extra "plotting" module that can provide simple utility plotting functions to make visualization easier for new users. In particular it should probably make use of seaborn's new ``relplot``, and provide an easy interface to datashader for large-scale data.
