### 2018-08-22 lmcinnes/umap#128 @lmcinnes
Documentation on sklearn pipelines
As per @terrytangyuan in openjournals/joss-reviews#861 , it would be beneficial if there was at least a brief section in the documentation outlining the use of UMAP with sklearn pipelines, and demonstrating using it with, for example, a clustering method, or a classifier, and potentially using GridSearchCV for hyper-parameter optimization.
