### 2018-08-22 #861 @terrytangyuan
@arokem Any suggestions regarding DOI reference?

@lmcinnes Thanks for your quick action. I don't think you have to do it for this review but I am glad that you are keeping track of them through issues.
