### 2018-08-25 8c9bd92
Fix spelling of russellrao; List all metrics
* Updated the list of metrics based on the ValueError produced when
calling an incorrect error.
