### 2018-08-30 #861 @arokem
Sorry for slowness here.

I think that the arXiv reference is sufficient.

Looks like that was the last item here? Is anything else required before this paper can be accepted?
