### 2018-08-30 #861 @terrytangyuan
Yep just checked all the boxes. We can move forward with this paper.
