### 2018-08-31 #861 @arokem
Wonderful! Congratulations @lmcinnes! 🎉

@lmcinnes : could you please create an archive for the accepted version (e.g., using https://zenodo.org/)? Once you have that send over the DOI as a comment on this issue, we will move ahead with the acceptance and paper.
