### 2018-09-02 #861 @arokem
@arfon : I believe this paper is ready to be published.

Nice work everyone! 🚀
